+++ packages/public/default-plugin-toolbar/src/plugin-toolbar-overlay-button.tsx
@@ -77,7 +77,11 @@ function WrappedModal({
           backgroundColor: '#ffffff',
           width: '90%',
           maxWidth: '600px',
-          inset: 'auto',
+          maxHeight: 'calc(100vh - 80px)',
+          top: 'auto',
+          left: 'auto',
+          right: 'auto',
+          bottom: 'auto',
           margin: '0 auto',
         },
       }}